feat: ui tweaks for editor toolbar order and look
