+++ src/components/Editor/EditorToolbar.module.css
@@ -1,8 +1,10 @@
 .toolbar {
-    position: relative;
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    padding: 4px;
-    border-block-end: 1px solid rgba(0, 0, 0, 0.1);
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 4px;
+  padding: 4px 8px;
+  border-block-end: 1px solid rgba(0, 0, 0, 0.1);
+  background-color: #f6f7fb;
 }
